requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
passlib[bcrypt]
bcrypt==3.2.0
pyjwt
python-multipart
python-dotenv
scikit-learn
pandas
numpy